api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pillow==10.2.0
torch==2.2.2
torchvision==0.17.2

